updated admin and user selection file
